read file with GITHUB WORKSPACE path

+++ src/index.ts
@@ -1,6 +1,7 @@
 import * as core from "@actions/core";
 import * as toml from "@iarna/toml";
 import * as fs from "fs";
+import * as path from "path";
 
 interface Jsonmap {
   [key: string]: any;
@@ -11,7 +12,11 @@ function run() {
     const fileName = core.getInput("file", { required: true });
     const key = core.getInput("key", { required: true });
     const value = core.getInput("value", { required: true });
-    let f = fs.readFileSync(fileName, "utf8");
+    const filePath = path.join(process.env.GITHUB_WORKSPACE, fileName);
+    if (!fs.existsSync(filePath)) {
+      throw new Error(`The toml file does not exist: ${fileName}`);
+    }
+    let f = fs.readFileSync(filePath, "utf8");
     let data = toml.parse(f);
     fs.writeFileSync(fileName, toml.stringify(setTomlByKey(data, key, value)));
   } catch (error) {}
